backend/requirements.txt
asgiref==3.11.1
Django==6.0.4
django-cors-headers==4.9.0
django-filter==25.1
djangorestframework==3.17.1
psycopg==3.3.3
psycopg-binary==3.3.3
psycopg2-binary==2.9.11
python-decouple==3.8
python-dotenv==1.2.2
sqlparse==0.5.5
